--- game/packages/cli/__init__.py
@@ -1,4 +1,0 @@
-#__init__.py
-#from .skuska import 
-from cli.cli_cards import *
-from cli.cli import *
