--- game/packages/dealer/__init__.py
@@ -1,8 +1,0 @@
-# __init__.py 
-from dealer import *
-from playerControl import *
-from cardControl import *
-
-# needed libraries
-from .. import player 
-
