+++ game/poker/__init__.py
@@ -1,0 +1,2 @@
+from poker.dealer import *
+from poker.detector import detector
