+++ game/poker/dealer/__init__.py
@@ -1,0 +1,8 @@
+# __init__.py 
+from dealer.dealer import *
+from dealer.playerControl import *
+from dealer.cardControl import *
+
+# dependancies
+from dealer.player import player
+
