+++ game/poker/dealer/dealer.py
@@ -10,7 +10,7 @@ Description: We need to create a "Dealer" for a game so we can
 """
 
 
-import player
+#import player
 import random
