+++ game/userinterface/__init__.py
@@ -1,0 +1,2 @@
+from cli import *
+from gui import *
