+++ game/userinterface/cli/__init__.py
@@ -1,0 +1,5 @@
+#__init__.py
+#from .skuska import 
+from cli.cards import cards
+from cli.questions import *
+from cli.printouts import *
